Address review feedback
* Avoid copy-pasted code for Scala 2 macros.
