+++ website/blog/2021-01-05-macromix.md
@@ -178,6 +178,7 @@ This feature is new and has some limitations that's good to be aware of:
   implemented as a Scala 2 macro.
 - Your Scala 3 library needs a compile-time dependecy on `scala-reflect:2.13.x`,
 
-Nevertheless, it's a phenomenal achievement that it's at all possible to publish
-Scala 2 and Scala 3 macros together. For certain libraries, I'm optimistic this
-feature will be a critical component to smoothen the Scala 3 transition.
+Given these limitations, MUnit will continue to publish for 2.13. Nevertheless,
+it's a phenomenal achievement that it's at all possible to publish Scala 2 and
+Scala 3 macros together. For certain libraries, I'm optimistic this feature will
+be a critical component to smoothen the Scala 3 transition.
